Adds Readline support to some apps
Adds ~C-u~ to delete line.  However I realized that macOS provides an OK
way of doing this i.e. ~C-a~ to move point to the beginning of the line,
and ~C-k~ to remove all text to the right.  I am still checking in this
file as a basis, but probably am going to disable loading it later.