Add toggle to ControlTapToEscape
Sadly I have to add this toggle because I was not able to get the
automatic disabling/enabling to occur for the NoMachine application.
The issue is with the NoMachine application using multiple windows, and
having a very strange focus/unfocus window behavior.  I had to give up
and provide this manual toggle to quickly disable the ControlTapToEscape
since having it enabled results in the Shift key not working inside
NoMachine windows.